+++ extensions/bitwarden/src/components/searchVault/utils/useVaultCaching.ts
@@ -1,16 +1,12 @@
 import { getPreferenceValues } from "@raycast/api";
 import { prepareFoldersForCache, prepareItemsForCache } from "~/components/searchVault/utils/caching";
+import { CACHE_KEYS } from "~/constants/general";
 import { Preferences } from "~/types/preferences";
 import { Folder, Item, Vault } from "~/types/vault";
 import { Cache } from "~/utils/cache";
 import { captureException } from "~/utils/development";
 import { useContentEncryptor } from "~/utils/hooks/useContentEncryptor";
 import useOnceEffect from "~/utils/hooks/useOnceEffect";
-
-export const CACHE_KEYS = {
-  IV: "iv",
-  VAULT: "vault",
-};
 
 function useVaultCaching() {
   const { encrypt, decrypt } = useContentEncryptor();