+++ extensions/docsearch/src/ipfs.tsx
@@ -1,5 +1,5 @@
 import { SearchDocumentation } from "./components";
 
 export default function Command(props: { arguments: { search?: string } }) {
-  return <SearchDocumentation docsName="IPFS" quickSearch={props.arguments?.search} />;
+  return <SearchDocumentation id="4e3fd0d3-d77e-4b13-91df-96cb2e36737d" quickSearch={props.arguments?.search} />;
 }