+++ extensions/harvest/src/stopTimer.tsx
@@ -1,4 +1,4 @@
-import { showToast, Toast, showHUD } from "@raycast/api";
+import { showToast, Toast, showHUD, open, launchCommand, LaunchType } from "@raycast/api";
 import { stopTimer } from "./services/harvest";
 
 export default async function main() {