+++ extensions/reflect/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Reflect Changelog
 
+## [OAuth & Append] - 2023-05-15
+
+### Added
+- Support for authenticating via OAuth
+- Append Command
+
 ## [Quick Append] - 2023-05-08
 
 ### New Commannds